Split netcdf integration tests into multiple modules.
